+++ soridam-api/src/main/java/sorisoop/soridam/api/noise/presentation/response/NoisePersistResponse.java
@@ -4,7 +4,7 @@ import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.REQUIRED;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
-import sorisoop.soridam.domain.place_service.noise.domain.Noise;
+import sorisoop.soridam.domain.place.noise.domain.Noise;
 
 @Builder
 public record NoisePersistResponse(
