+++ soridam-api/src/main/java/sorisoop/soridam/api/place/presentation/response/PlaceResponse.java
@@ -5,8 +5,8 @@ import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.REQUIRED;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
-import sorisoop.soridam.domain.place_service.place.domain.Place;
-import sorisoop.soridam.domain.place_service.place.domain.enums.Category;
+import sorisoop.soridam.domain.place.place.domain.Place;
+import sorisoop.soridam.domain.place.place.domain.enums.Category;
 
 @Builder
 public record PlaceResponse(
